+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,25 +1,25 @@
 object Versions {
     private val major = 0
     private val minor = 4
-    private val patch = 26
+    private val patch = 27
     private val semantic: String = ""//""-alpha.2"
 
     val BidonVersionName = mainVersion + semanticVersion
 
     object Adapters {
-        val Admob = "$mainVersion.1" + semanticVersion
-        val Applovin = "$mainVersion.1" + semanticVersion
-        val BidMachine = "$mainVersion.1" + semanticVersion
-        val DTExchange = "$mainVersion.1" + semanticVersion
+        val Admob = "$mainVersion.0" + semanticVersion
+        val Applovin = "$mainVersion.0" + semanticVersion
+        val BidMachine = "$mainVersion.0" + semanticVersion
+        val DTExchange = "$mainVersion.0" + semanticVersion
         val UnityAds = "$mainVersion.0" + semanticVersion
-        val BigoAds = "$mainVersion.1" + semanticVersion
-        val Mintegral = "$mainVersion.1" + semanticVersion
-        val Vungle = "$mainVersion.1" + semanticVersion
+        val BigoAds = "$mainVersion.0" + semanticVersion
+        val Mintegral = "$mainVersion.0" + semanticVersion
+        val Vungle = "$mainVersion.0" + semanticVersion
         val Meta = "$mainVersion.0" + semanticVersion
-        val Inmobi = "$mainVersion.1" + semanticVersion
+        val Inmobi = "$mainVersion.0" + semanticVersion
         val Amazon = "$mainVersion.0" + semanticVersion
         val MobileFuse = "$mainVersion.0" + semanticVersion
-        val Gam = "$mainVersion.1" + semanticVersion
+        val Gam = "$mainVersion.0" + semanticVersion
 
         val IronSource = "$mainVersion.0"
         val Appsflyer = "$mainVersion.0"